rename file to useWishlist.js
