+++ packages/pwa/app/pages/account/wishlist/index.jsx
@@ -10,7 +10,7 @@ import {FormattedMessage, useIntl} from 'react-intl'
 import PageActionPlaceHolder from '../../../components/page-action-placeholder'
 import {WishlistIcon} from '../../../components/icons'
 import useNavigation from '../../../hooks/use-navigation'
-import useWishlist from '../../../commerce-api/hooks/useCustomerProductLists'
+import useWishlist from '../../../commerce-api/hooks/useWishlist'
 import {Box, Flex, Skeleton} from '@chakra-ui/react'
 import {API_ERROR_MESSAGE, customerProductListTypes} from '../../../constants'
 import ProductItem from '../../../components/product-item/index'
