+++ packages/pwa/app/pages/account/wishlist/partials/wishlist-secondary-button-group.test.js
@@ -21,9 +21,9 @@ const MockedComponent = () => {
     )
 }
 
-jest.mock('../../../../commerce-api/hooks/useCustomerProductLists', () => {
+jest.mock('../../../../commerce-api/hooks/useWishlist', () => {
     const originalModule = jest.requireActual(
-        '../../../../commerce-api/hooks/useCustomerProductLists'
+        '../../../../commerce-api/hooks/useWishlist'
     )
     const useCustomerProductLists = originalModule.default
     return () => {
