+++ packages/pwa/app/pages/checkout/index.test.js
@@ -54,8 +54,8 @@ jest.mock('../../commerce-api/pkce', () => {
     }
 })
 
-jest.mock('../../commerce-api/hooks/useCustomerProductLists', () => {
-    const originalModule = jest.requireActual('../../commerce-api/hooks/useCustomerProductLists')
+jest.mock('../../commerce-api/hooks/useWishlist', () => {
+    const originalModule = jest.requireActual('../../commerce-api/hooks/useWishlist')
     const useCustomerProductLists = originalModule.default
 
     return () => {
